Major checkpoint

+++ app/controllers/students_controller.rb
@@ -1,5 +1,10 @@
 class StudentsController < ApplicationController
   protect_from_forgery
+  layout "main"
+  
+  def site_section
+  	:students_section
+  end
 
   def index
     @semester = Semester.find(params[:semester_id])
@@ -13,7 +18,11 @@ class StudentsController < ApplicationController
   end
   
   def show
-  	redirect_to edit_student_path
+    @student = Student.find(params[:id])
+    return unless valid_student?(@student)
+
+    @semester = params.include?(:semester_id) ? Semester.find(params[:semester_id]) : @student.semester
+    return unless valid_semester?(@semester)
   end
 
   def new
@@ -21,6 +30,8 @@ class StudentsController < ApplicationController
     return unless valid_semester?(@semester)
     
     @student = flash.key?(:student) ? Student.new(flash[:student]) : Student.new
+    
+    @teachers = @semester.teachers
   end
 
   def create
@@ -29,13 +40,14 @@ class StudentsController < ApplicationController
 
     @student = @semester.students.build(params[:student])
     
-    @teacher = Teacher.find(params[:teacher])
+    @teacher = params[:teacher] && !params[:teacher].empty? ? Teacher.find(params[:teacher]) : nil
     if @teacher
       @student.teacher = @teacher
     else
-      flash[:warning] = [[:teacher,"A valid teacher was not selected."]]
+      flash[:warning] = [[:teacher,"Must select a valid teacher."]]
       flash[:student] = params[:student] # Save fields so the user doesn't have to re-enter everything again
-      redirect_to edit_student_path
+      redirect_to new_semester_student_path(@semester)
+      return
     end
 
     if @student.save
@@ -53,14 +65,6 @@ class StudentsController < ApplicationController
 
     @semester = params.include?(:semester_id) ? Semester.find(params[:semester_id]) : @student.semester
     return unless valid_semester?(@semester)
-
-    @classes = @semester.courses
-    @classes.unshift(Course.new(:name => "Select Item:"))
-    @classes.each do |course|
-      course.modify_name_for_enrollments(@student.id)
-    end     
-
-    @enrollment = flash.key?(:enrollment) ? Enrollment.new(flash[:enrollment]) : Enrollment.new
   end
 
   def update
@@ -82,7 +86,7 @@ class StudentsController < ApplicationController
     end
 
     if @student.save 
-      redirect_to semester_students_path(@semester), :notice => "#{@student.first_name} #{@student.last_name}'s information was successfully updated."
+      redirect_to student_path(@student), :notice => "#{@student.first_name} #{@student.last_name}'s information was successfully updated."
     else
       flash[:warning] = @student.errors
       flash[:student] = params[:student] # Save fields so the user doesn't have to re-enter everything again
@@ -106,19 +110,88 @@ class StudentsController < ApplicationController
     redirect_to semester_students_path(@semester)
   end
   
-  def enroll
+  def index_enrollments
     @student = Student.find(params[:id])
     return unless valid_student?(@student)
 
     @semester = params.include?(:semester_id) ? Semester.find(params[:semester_id]) : @student.semester
     return unless valid_semester?(@semester)
 
-    @enrollment = @student.find_enrollment(params[:enrollment][:course_id])
+  	render 'students/enrollments/index'
+  end
+  
+  def show_enrollment
+    @student = Student.find(params[:id])
+    return unless valid_student?(@student)
+
+    @course = Course.find(params[:course_id])
+    return unless valid_course?(@course)
+
+    @semester = params.include?(:semester_id) ? Semester.find(params[:semester_id]) : @course.semester
+    return unless valid_semester?(@semester)
+    
+    @enrollment = @student.find_enrollment(@course.id)
+    return unless valid_enrollment?(@enrollment)
+ 
+  	render 'students/enrollments/show'
+  end
+  
+  def new_enrollment
+    @student = Student.find(params[:id])
+    return unless valid_student?(@student)
+
+    @semester = params.include?(:semester_id) ? Semester.find(params[:semester_id]) : @student.semester
+    return unless valid_semester?(@semester)
+
+    @classes = @semester.courses
+    @classes.unshift(Course.new(:name => "Select Class"))
+    @classes.each do |course|
+      course.modify_name_for_enrollments(@student.id)
+    end
+
+    @enrollment = flash.key?(:enrollment) ? Enrollment.new(flash[:enrollment]) : Enrollment.new
+
+  	render 'students/enrollments/new'
+  end
+  
+  def edit_enrollment
+    @student = Student.find(params[:id])
+    return unless valid_student?(@student)
+
+    @course = Course.find(params[:course_id])
+    return unless valid_course?(@course)
+
+    @semester = params.include?(:semester_id) ? Semester.find(params[:semester_id]) : @course.semester
+    return unless valid_semester?(@semester)
+    
+    @enrollment = @student.find_enrollment(@course.id)
+    return unless valid_enrollment?(@enrollment)
+
+    @classes = @semester.courses
+    @classes.unshift(Course.new(:name => "Select Class"))
+    @classes.each do |course|
+      course.modify_name_for_enrollments(@student.id)
+    end   
+
+  	render 'students/enrollments/edit'
+  end
+  
+  def create_enrollment
+    @student = Student.find(params[:id])
+    return unless valid_student?(@student)
+    
+    @course = params[:enrollment][:course_id].present? ? Course.find(params[:enrollment][:course_id]) : nil
+    return unless valid_course?(@course, student_path(@student))
+
+    @semester = params.include?(:semester_id) ? Semester.find(params[:semester_id]) : @course.semester
+    return unless valid_semester?(@semester)
+
+    @enrollment = @student.find_enrollment(@course.id)
     if @enrollment.nil?
       # Student not already enrolled, enroll them
       @enrollment = @student.create_enrollment(params[:enrollment])
       if not @enrollment.new_record?
-        flash[:notice] = "#{@student.first_name} #{@student.last_name} was successfully enrolled in #{@enrollment.course.name}."
+        flash[:notice] = "#{@student.first_name} #{@student.last_name} was successfully enrolled in #{@course.name}."
       else
         flash[:warning] = @enrollment.errors
         flash[:enrollment] = params[:enrollment]
@@ -126,35 +199,34 @@ class StudentsController < ApplicationController
     else
       # Student already enrolled, just update the enrollment record
       if @enrollment.update_attributes(params[:enrollment])
-        flash[:notice] = "#{@student.first_name} #{@student.last_name}'s enrollment in #{@enrollment.course.name} was successfully updated."
+        flash[:notice] = "#{@student.first_name} #{@student.last_name}'s enrollment in #{@course.name} was successfully updated."
       else
         flash[:warning] = @enrollment.errors
         flash[:enrollment] = params[:enrollment]
       end
     end
-    redirect_to edit_student_path
+    redirect_to student_path
   end
   
-  def unenroll
+  def destroy_enrollment
     @student = Student.find_by_id params[:id]
     return unless valid_student?(@student)
     
-    @enrollment = @student.find_enrollment(params[:course_id])
+    @course = Course.find(params[:course_id])
+    return unless valid_course?(@course)
+
+    @enrollment = @student.find_enrollment(@course.id)
     if @enrollment.nil?
       flash[:warning] = [[:enrollment, "Couldn't unenroll, the given student is not enrolled for the given course."]]
       redirect_to edit_student_path
       return
     end
 
-    first_name = @student.first_name
-    last_name = @student.last_name
-    course_name = @enrollment.course.name
-
     if @enrollment.destroy
-      flash[:notice] = "Successfully deleted #{first_name} #{last_name}'s enrollment in #{course_name}"
+      flash[:notice] = "#{@student.first_name} #{@student.last_name} was successfully unenrolled from #{@course.name}"
     else
       flash[:warning] = @enrollment.errors
     end
-    redirect_to edit_student_path
+    redirect_to student_path
   end
 end
